services/python-worker/requirements.txt
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.21.0
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
onnxruntime>=1.12.0
pillow
pyclipper
shapely
shapely
vietocr
torch>=1.13.0
torchvision>=0.14.0
